+++ NEWS.md
@@ -1,8 +1,13 @@
 Version 1.9.0.9000
 ==================
 
+### NEW FEATURES
+
+* `as_datetime()` accepts multiple formats in format argument, just like `as_date()` does.
+
 ### BUG FIXES
 
+* [#1097](https://github.com/tidyverse/lubridate/issues/1097) Fix `as_date("character")` to work correctly with formats that include extra characters.
 * [#1098](https://github.com/tidyverse/lubridate/issues/1098) Roll over the month boundary in `make_dateime()` when units exceed their maximal values.
 * [#1090](https://github.com/tidyverse/lubridate/issues/1090) timechange has been moved from Depends to Imports.